% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U1.R
\docType{data}
\name{U1}
\alias{U1}
\title{Population U1}
\format{
A vector 430x1:
}
\usage{
U1
}
\description{
Australian cattle farms stratified by industrial regions
}
\details{
U1
}
\references{
1. Chambers, R., Dunstan, R., 1986.
Estimating distribution functions from survey data. Biometrika 73, 3, 597–604.
}
\keyword{datasets}
