% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U15.R
\docType{data}
\name{U15}
\alias{U15}
\title{Population U15}
\format{
A vector 284x1:
}
\usage{
U15
}
\description{
Population in thousands of 284 municipalities in Sweden in 1975
}
\details{
U15
}
\references{
Särndal, C.E., Swensson,B., Wretman, J. (2003).
Model Assisted Survey Sampling, 1st edition, Springer.
}
\keyword{datasets}
