% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U21.R
\docType{data}
\name{U21}
\alias{U21}
\title{Population U21}
\format{
A vector 357x1:
}
\usage{
U21
}
\description{
Million dollar funds from major US commercial banks
}
\details{
U21
}
\references{
Särndal, C.E., Swensson,B., Wretman, J. (2003).
Model Assisted Survey Sampling, 1st edition, Springer.
}
\keyword{datasets}
